refacto infra code split into different files
